requirements.txt
beautifulsoup4==4.9.1
requests==2.23.0
python-telegram-bot==13.11
hanziconv==0.3.2
python-dotenv==0.15.0
schedule==1.0.0
pyyaml==5.4.1
